+++ TableViewKit/NibClassType.swift
@@ -7,10 +7,17 @@ public enum NibClassType<T> {
     case nib(UINib, T.Type)
     /// If it must be loaded from a Class
     case `class`(T.Type)
+    /// If it must be loaded from a Storyboard prototype
+    case prototype(String, T.Type)
 
     /// The reusable identifier for the type
     public var reusableIdentifier: String {
-        return String(describing: typeClass)
+        switch self {
+        case .class, .nib:
+            return String(describing: typeClass)
+        case .prototype(let identifier, _):
+            return identifier
+        }
     }
 
     /// The type class
@@ -20,8 +27,11 @@ public enum NibClassType<T> {
             return cellClass
         case .nib(_, let cellClass):
             return cellClass
+        case .prototype(_, let cellClass):
+            return cellClass
         }
     }
+
 }
 
 extension NibClassType where T: UITableViewCell {
@@ -31,6 +41,8 @@ extension NibClassType where T: UITableViewCell {
             return NibClassType<UITableViewCell>.class(type)
         case .nib(let nib, let type):
             return NibClassType<UITableViewCell>.nib(nib, type)
+        case .prototype(let identifier, let type):
+            return NibClassType<UITableViewCell>.prototype(identifier, type)
         }
     }
 }
@@ -42,6 +54,8 @@ extension NibClassType where T: UITableViewHeaderFooterView {
             return NibClassType<UITableViewHeaderFooterView>.class(type)
         case .nib(let nib, let type):
             return NibClassType<UITableViewHeaderFooterView>.nib(nib, type)
+        case .prototype(let identifier, let type):
+            return NibClassType<UITableViewHeaderFooterView>.prototype(identifier, type)
         }
     }
 }